Move Permission sync code to StartupRunner class
